refactor: Abstract Home component logic to custom hooks and utils
- Extract state management and event handlers from Home.tsx into useHome custom hook.
- Move utility functions like checkIfMobile to utils for shared access.
- Enhance code readability and maintainability by modularizing complex component logic.
